requirements.txt
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
yfinance==0.1.55
Pyrebase4==4.4.3
mysqlclient==2.0.1
django-crontab==0.7.1
PyMySQL==1.0.2
python-decouple==3.4